more exploring of underflow and overflow

+++ 001-counter/program/src/state.rs
@@ -2,7 +2,7 @@ use borsh::{BorshDeserialize, BorshSerialize};
 
 #[derive(BorshSerialize, BorshDeserialize, Debug, PartialEq)]
 pub struct Counter {
-    pub value: u32,
+    pub value: u16,
 }
 
 impl Default for Counter {
@@ -15,7 +15,6 @@ impl Default for Counter {
 mod tests {
     use crate::state::Counter;
     use borsh::{BorshDeserialize, BorshSerialize};
-    // use std::mem::size_of;
 
     #[test]
     fn test_counter_serialization() {
@@ -23,7 +22,5 @@ mod tests {
         let encoded = original.try_to_vec().expect("Could not encode");
         let decoded = Counter::try_from_slice(&encoded).expect("Could not decode");
         assert_eq!(original, decoded);
-
-        // println!("size_of Counter {}", size_of::<Counter>());
     }
 }
